requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
python-dotenv==0.13.0
psycopg2-binary==2.8.5
uwsgi==2.0.18
requests==2.23.0 
werkzeug==2.0.3
itsdangerous==2.0.1
jinja2==3.0.0